Moved more tie functions to C++.
Also removed some redundent code and comments from the VB.

+++ Library/DBLibrary/mathhelp.cpp
@@ -9,13 +9,13 @@ float Angle(float x1, float y1, float x2, float y2)
 	if (dx == 0)
 	{
 		if (dy >= 0)
-			return M_PI_2 * 3;
-		return M_PI_2;
+			return (float)M_PI_2 * 3;
+		return (float)M_PI_2;
 	}
 	else {
 		if (dx < 0)
 			return atan(dy / dx);
-		return atan(dy / dx) + M_PI;
+		return atan(dy / dx) + (float)M_PI;
 	}
 }
 
@@ -23,18 +23,18 @@ float AngleDifference(float a1, float a2) {
 	float diff = a1 - a2;
 
 	if (diff > M_PI)
-		diff = (-2 * M_PI - diff);
+		diff = (-2 * (float)M_PI - diff);
 	if (diff < -M_PI)
-		diff += 2 * M_PI;
+		diff += 2 * (float)M_PI;
 
 	return diff;
 }
 
 float AngleNormalise(float a) {
 	while (a < 0)
-		a += M_PI * 2;
+		a += (float)M_PI * 2;
 	while (a > M_PI * 2)
-		a -= M_PI * 2;
+		a -= (float)M_PI * 2;
 
 	return a;
 }
